requirements.txt
torch
torchvision
pillow
gradio
ultralytics